Add `TestSetCategory` enumerators, allow choosing category for new quiz, indicate started quiz
Additionally, the `ComposeLoadingPrompt` function was moved to the `util` package, because multiple activities will be using that, examples being  `MainActivity` and `QuizActivity` currently.

+++ app/src/main/java/com/provigz/avtotest/db/entity/TestSet.kt
@@ -32,7 +32,7 @@ data class TestSet(
 
     constructor(model: com.provigz.avtotest.model.TestSet) : this(
         id = model.id,
-        categoryID = model.subCategory.id,
+        categoryID = model.subCategory.id.toInt(),
         categoryName = model.subCategory.name,
         durationMinutes = model.subCategory.durationMinutes,
         questionIDs = emptyList()
